pools: Pool page purchase tokens in place
