+++ packages/abacus-pools/src/sections/CreatePool/CreatePool.tsx
@@ -20,6 +20,7 @@ const Container = styled(LayoutContainer)`
 const StyledSection = styled(Section)<{ complete: boolean }>`
   display: flex;
   flex-direction: column;
+  gap: 24px;
 
   ${({ complete }) => (complete ? "" : Media.sm`padding: 20px 76px;`)}
 `
